Further progress on ME5
Area errors still problematic - I've tried to remove the errors that were flagged up in the code but weren't causing a compiling error but they still appear.  A bigger issue is the pdf not compiling when you try to shade in certain areas within the scrap - for instance the large boulder piles halfway down.